requirements.txt
langchain
chromadb
beautifulsoup4
pymupdf
networkx
streamlit
requests
fastapi
uvicorn
python-dotenv
tenacity
celery
redis
sqlalchemy
openai
langchain-llama
langchain-openai